move existing fb generator logic to generator
