+++ filebeat/generator/fields/fields_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package main
+package fields
 
 import (
 	"reflect"
